+++ src/test/java/io/github/astrapi69/mystic/crypt/key/PublicKeyStringEncryptorTest.java
@@ -32,35 +32,37 @@ import java.security.PublicKey;
 import org.testng.annotations.Test;
 
 import io.github.astrapi69.crypt.data.key.reader.PublicKeyReader;
+import io.github.astrapi69.file.read.ReadFileExtensions;
 import io.github.astrapi69.file.search.PathFinder;
 
-/**
- * The unit test class for the class {@link PublicKeyHexEncryptor}
- */
-public class PublicKeyHexEncryptorTest
+public class PublicKeyStringEncryptorTest
 {
 
 	/**
-	 * Test method for {@link PublicKeyHexEncryptor} constructors
+	 * Test to encrypt with {@link PublicKeyStringEncryptor#encrypt(String)}
 	 *
 	 * @throws Exception
-	 *             is thrown if any error occurs
+	 *             is thrown if the encryption or the decryption fails
 	 */
-	@Test
-	public void testConstructors() throws Exception
+	@Test(enabled = true)
+	public void testEncrypt() throws Exception
 	{
+		PublicKeyStringEncryptor encryptor;
 		PublicKey publicKey;
 		File publickeyDerDir;
 		File publickeyDerFile;
-		PublicKeyHexEncryptor encryptor;
-		String encrypted;
+		File xml;
+		byte[] encrypted;
 
 		publickeyDerDir = new File(PathFinder.getSrcTestResourcesDir(), "der");
 		publickeyDerFile = new File(publickeyDerDir, "public.der");
 		publicKey = PublicKeyReader.readPublicKey(publickeyDerFile);
-		encryptor = new PublicKeyHexEncryptor(publicKey);
+		xml = PathFinder.getRelativePath(PathFinder.getSrcTestResourcesDir(), "xml", "company.xml");
+		String xmlString = ReadFileExtensions.fromFile(xml);
+
+		encryptor = new PublicKeyStringEncryptor(publicKey);
 		assertNotNull(encryptor);
-		encrypted = encryptor.encrypt("foo");
+		encrypted = encryptor.encrypt(xmlString);
 		assertNotNull(encrypted);
 	}
 }